requirements.txt
bs4
pymongo
pypokedex
python-telegram-bot==13.14
